requirements.txt
# HTTP requests & scraping libraries
requests
beautifulsoup4
lxml
pandas

# Optional: type checking & data models
pydantic

# Logging & pretty printing
loguru

# Optional: async scraping
httpx
aiohttp

# Browser automation & advanced scraping
playwright

# Database integration
sqlalchemy
psycopg2-binary
pydantic
alembic

python-dotenv
